+++ src/lift/lifter/magic.rs
@@ -53,6 +53,7 @@ impl Lifter {
             body,
             foreign: false,
             generic_ret_from_param: None,
+            private_window: None,
             span: SP,
         })))
     }